+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Added
+
+- Hot-reload on SLI plugin file loader, triggered by HTTP webhook.
+- Added `hot-reload-addr` flag with the hot reload http server address.
+- Added `hot-reload-path` flag with the hot reload http server webhookpath webhook.
+
 ### Changed
 
 - (Internal) SLI Plugins are retrieved from a repository service instead of getting them from a `map`.